added plot_tiled_data.m
plot_tiled_data.m takes an n-by-t matrix and plots each column (t) in a different subplot tile. This is useful when plotting multiple channels from one participant or multiple participants from one sample, if you don't want the data to crowd in a single graph. The tiles are supposed to be optimally distributed based on the dimensions of your monitor, but no guarantees!

plot_tiled_data takes arguments:
- data_array (the n-by-t matrix)
- type (plot, dot, bar, hist, or imagesc)
- titletext (a string that will be enumerated for each subplot title, e.g., "Subject #")
- xlabtext (a string that will appear below each x-axis, e.g., "Time")